+++ package.json
@@ -114,7 +114,7 @@
     "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-react": "7.24.0",
     "eslint-plugin-react-hooks": "4.2.0",
-    "flow-bin": "0.89.0",
+    "flow-bin": "0.156.0",
     "jasmine-check": "0.1.5",
     "jest": "26.6.3",
     "make-synchronous": "^0.1.1",
